Move root-level JS project files into `munkey` dir
