--- config.json
@@ -1,5 +1,0 @@
-{
-  "couchdb": {
-    "uuid": "289739b0-0d33-4870-84c1-9267d65c0209"
-  }
-}
