--- localhost/csr.pem
@@ -1,16 +1,0 @@
------BEGIN CERTIFICATE REQUEST-----
-MIIChzCCAW8CAQAwQjELMAkGA1UEBhMCVVMxFTATBgNVBAcMDERlZmF1bHQgQ2l0
-eTEcMBoGA1UECgwTRGVmYXVsdCBDb21wYW55IEx0ZDCCASIwDQYJKoZIhvcNAQEB
-BQADggEPADCCAQoCggEBAMxePK+2UJJIu37foPBWSOfWy2BVfrqIkEMJ8rCBA0tC
-D2GLu7Pp23NN/N1aSfFWKSLBlE9ipxi6x13dTEqj2GO4sq9ONAU87F6C1xuNz3qU
-VOrmKQFFfjn2Wi0uULfaHqDUH0JNe1sZBUhXKXAXA3pHznvYjqixxL0QHlEO+knG
-yB3fxtPDsVAz1pSCo48HA4gF5HyYypgbqtZvKCW0uLYcNHrigM/K6yhffYrssFGp
-B2Tu+CK8Kem6iGgaQVBM/Y4+Dph/83YbaYBGQXyEfGLfOaNZ7EalGU15eCThox1w
-mGDKxo31mihQHVVQkRwEOu7ZdAnrnIIqhg1MfbfZGXsCAwEAAaAAMA0GCSqGSIb3
-DQEBCwUAA4IBAQBhQ55pSEvnyIfb5MbjvDqSSwn1+rNZwZc8OQ4nwA4S8NKmKSFT
-qtLPa+s5C2ML6cdz0nWOuVgRrPmGTC6T3Wb8vQcrwIBXVwtBJoLz9qK7xg7hJGME
-3bSHNv5vZgki6JNM+qyR/UAWMaoe1WvXOtdmm0rH16ye4sSKaKFrHdsyFD5KxlpE
-qTCu4/bSSrzPuK6LGDQn1eFiJmS+uOQHfdkVORvkHp535E6a9Ybrsr6WiTByv5bc
-zJi4KbV6QtYYjASQb79qF5K02+8Lci7bb02XoxTObZWBSR1iDfwjv8F4BE/+CeNW
-RpWPKG6kIwApVUS9IlrlwFiwduCubFM+tceY
------END CERTIFICATE REQUEST-----
