--- localhost/tls.crt
@@ -1,19 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDCzCCAfMCFH6KzlAt2udwaL20C0EwIpsZILf3MA0GCSqGSIb3DQEBCwUAMEIx
-CzAJBgNVBAYTAlVTMRUwEwYDVQQHDAxEZWZhdWx0IENpdHkxHDAaBgNVBAoME0Rl
-ZmF1bHQgQ29tcGFueSBMdGQwHhcNMjExMTI1MTkyMjIwWhcNNDkwNDExMTkyMjIw
-WjBCMQswCQYDVQQGEwJVUzEVMBMGA1UEBwwMRGVmYXVsdCBDaXR5MRwwGgYDVQQK
-DBNEZWZhdWx0IENvbXBhbnkgTHRkMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIB
-CgKCAQEAzF48r7ZQkki7ft+g8FZI59bLYFV+uoiQQwnysIEDS0IPYYu7s+nbc038
-3VpJ8VYpIsGUT2KnGLrHXd1MSqPYY7iyr040BTzsXoLXG43PepRU6uYpAUV+OfZa
-LS5Qt9oeoNQfQk17WxkFSFcpcBcDekfOe9iOqLHEvRAeUQ76ScbIHd/G08OxUDPW
-lIKjjwcDiAXkfJjKmBuq1m8oJbS4thw0euKAz8rrKF99iuywUakHZO74Irwp6bqI
-aBpBUEz9jj4OmH/zdhtpgEZBfIR8Yt85o1nsRqUZTXl4JOGjHXCYYMrGjfWaKFAd
-VVCRHAQ67tl0CeucgiqGDUx9t9kZewIDAQABMA0GCSqGSIb3DQEBCwUAA4IBAQA/
-5sHm2veq1OXtC1LrsEr8fwh9k6Xkr536HCdxOlj19dlYeIXfScBbY6mv8v7zuSks
-A83hQ9WgG8SvXCDhZiVapkKk+yZ88QrCvhMLyJse0r1NwPNLAVuZqnwQnSICKSU4
-fd7t8isUm3gnI+n/bDlYij3xeTAKAJb+LBuryrUWVXdW1CsmKJmUDRlD90EvtQrQ
-8o5q21QGVnRY46x9/rttHAKocxQk3U7lnqYB7iOLT025m80lpO4q67JUdw43XM/b
-89sZGgXS0HImqlJSCSHS0n+Cm/GC0B2zPBPUafY3qabSpgYZOKgCofbhA+ArKHn+
-LQxNM1QlUq/qzpSotE5D
------END CERTIFICATE-----
