--- localhost/tls.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpgIBAAKCAQEAzF48r7ZQkki7ft+g8FZI59bLYFV+uoiQQwnysIEDS0IPYYu7
-s+nbc0383VpJ8VYpIsGUT2KnGLrHXd1MSqPYY7iyr040BTzsXoLXG43PepRU6uYp
-AUV+OfZaLS5Qt9oeoNQfQk17WxkFSFcpcBcDekfOe9iOqLHEvRAeUQ76ScbIHd/G
-08OxUDPWlIKjjwcDiAXkfJjKmBuq1m8oJbS4thw0euKAz8rrKF99iuywUakHZO74
-Irwp6bqIaBpBUEz9jj4OmH/zdhtpgEZBfIR8Yt85o1nsRqUZTXl4JOGjHXCYYMrG
-jfWaKFAdVVCRHAQ67tl0CeucgiqGDUx9t9kZewIDAQABAoIBAQCDnq22/NQnYnBe
-5efg4bFSnyOch3N27zz58A49XtmgPotpZ3UcCiErwa55YQz+QV984u+BsSes5Z5A
-9aWM7LkQgIOUI+mc9f/FXr7rIAngCGgoYNNH3lnNOrwZHRsfTXssWXFIYl5v7U1Z
-qckmR6wVtOlnGbHHM7ZhjV/5FIxdtmG6MibxtkQSgm4T8954XMcziA86LnQ1v5M7
-7mHHyBm7Qe3ib0OdSqVz9O1blFd8BFEG9L9XaHkDhFXO5iCN2XBwl97tpcWQTFHA
-mf6dmyjtfBziV0SFErbUyJYDrNRx5UQeCVNOpfT5rpM9eA49xx7QxyVBkyqzLp/a
-pMvI5vkBAoGBAOmw4RBrFHYDgkPD7FHYp0avEGmV0JkR+vA+7W01BZr9Igmy6hxp
-3EmCNG/L2ZI55Ept/RYBOkDg+bi9bz3WvPYa0MeDOi6XEMsaPUdnWqWWA5icHl4v
-/KheUt/iTCibGYWX5Eu7PFdpDdmFzF9A1uTOHuamOZpM1sMOTwLLqHSFAoGBAN/g
-vnaH8weejwXFtN+D8LkPmUW1TdTrUiqbYPGJwiCdDhpB37TeLeMQj8bOAk98VKvR
-EB9jcw9j/WgaGSkzVvpi4Q2ZnGi9jNaa8wzEErQBnm3Mfw41LsYXt+ZnDg3UFrJ/
-dEBzqYSLao51+VoX/ZyhXzKyvAbT77EVedm/b7X/AoGBANGllyONjNuapkB5Agcj
-IF4vK8AtYOgR01e4fHPef1reAK1GzvQSnEduAfDRpiyitwV2yvf0vff6XM25VJTb
-ksYOpIJ4XbfyWmR688KdHBs1C6DbXfsNfdLmW97yO3SqQCkzbOHr5WRdoMkmWYSS
-vLajm+E7+q1MhdaTfZp6bnOpAoGBAMsdeU/K6giYp4QCKqa7avRLnbCr3FB3q5Vy
-YRLi/BhgxYG3EEJlbVZcGUWydFAvKha0V59St/pXqnn/a6KArMIAYdTX8BrrFlNC
-Q47qeVmNOnK9nOyD/crFjBhimVKcgHczwYIULdFON7/GcxN1PqgTlG5H0OWU9RtB
-s8qFr9F7AoGBAJC/rw2A58NQjv4lQEKHRxP1wkTH5kc7WqJS3kQ4HDNwsAe9YbN8
-SyagwqueZIEEimgXguCYIMnk9Z/gN4tdS2J8UZxDki2VUs2Q0b/M7TRb8MtImdox
-SrrNE80JiGaJN7Ys9rlys4kNDD9oXq6Svj4DR+201ykQii7NcdtaABq6
------END RSA PRIVATE KEY-----
