+++ munkey/test/integration/webServerOperations.test.ts
@@ -1,4 +1,4 @@
-import * as services from "../../munkey/services";
+import * as services from "../../services";
 import express from "express";
 import http from "http";
