+++ munkey/tsconfig.json
@@ -4,7 +4,6 @@
         "esModuleInterop": true,
         "moduleResolution": "node",
         "target": "ESNext",
-        "composite": true,
         "sourceMap": true,
         "types": [
             "node"
