--- tsconfig.json
@@ -1,21 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "esModuleInterop": true,
-        "types": [
-            "node"
-        ],
-        "target": "ES6",
-        "sourceMap": false,
-        "outDir": "./bin"
-    },
-    "references": [
-        {
-            "path": "./munkey"
-        }
-    ],
-    "files": [],
-    "exclude": [
-        "node_modules"
-    ]
-}
